% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{celladmix_add_corrected_assay}
\alias{celladmix_add_corrected_assay}
\title{Add Corrected cellAdmix Counts as a Seurat Assay}
\usage{
celladmix_add_corrected_assay(object, correction, assay = "cellAdmix")
}
\arguments{
\item{object}{A Seurat object.}

\item{correction}{A \code{CellAdmixCorrection} object.}

\item{assay}{Name of the assay to create.}
}
\value{
The Seurat object with a corrected sparse counts assay.
}
\description{
Add Corrected cellAdmix Counts as a Seurat Assay
}
