% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{celladmix_add_factors_to_seurat}
\alias{celladmix_add_factors_to_seurat}
\title{Add cellAdmix Factor Metadata to a Seurat Object}
\usage{
celladmix_add_factors_to_seurat(object, fit, prefix = "celladmix_")
}
\arguments{
\item{object}{A Seurat object.}

\item{fit}{A \code{CellAdmixFit} object.}

\item{prefix}{Metadata column prefix.}
}
\value{
The Seurat object with factor metadata columns added.
}
\description{
Add cellAdmix Factor Metadata to a Seurat Object
}
