% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\name{celladmix_domain_composition}
\alias{celladmix_domain_composition}
\title{Domain Annotation Composition}
\usage{
celladmix_domain_composition(domains)
}
\arguments{
\item{domains}{A \code{celladmix_domains} object.}
}
\value{
The per-domain annotation composition data frame.
}
\description{
Domain Annotation Composition
}
