% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\name{celladmix_domain_diagnostics}
\alias{celladmix_domain_diagnostics}
\title{Domain Identification Diagnostics}
\usage{
celladmix_domain_diagnostics(domains)
}
\arguments{
\item{domains}{A \code{celladmix_domains} object.}
}
\value{
The domain-identification diagnostics list.
}
\description{
Domain Identification Diagnostics
}
