% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\name{celladmix_plot_domain_composition}
\alias{celladmix_plot_domain_composition}
\title{Plot Domain Composition}
\usage{
celladmix_plot_domain_composition(domains)
}
\arguments{
\item{domains}{A \code{celladmix_domains} object.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Domain Composition
}
