% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\name{celladmix_plot_domain_umap}
\alias{celladmix_plot_domain_umap}
\title{Plot Domains on a Cell UMAP}
\usage{
celladmix_plot_domain_umap(domains, clust = NULL, color = "domain_label")
}
\arguments{
\item{domains}{A \code{celladmix_domains} object.}

\item{clust}{A \code{celladmix_clusters} object. Defaults to \code{domains$clust} when
available.}

\item{color}{Domain table column used for point color.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Domains on a Cell UMAP
}
