% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\name{celladmix_plot_domains_spatial}
\alias{celladmix_plot_domains_spatial}
\title{Plot Spatial Domain Assignments}
\usage{
celladmix_plot_domains_spatial(
  domains,
  color = "domain_label",
  point_size = 0.15,
  alpha = 0.7
)
}
\arguments{
\item{domains}{A \code{celladmix_domains} object or its \verb{$domains} data frame.}

\item{color}{Column used for point color.}

\item{point_size}{Point size.}

\item{alpha}{Point alpha.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Spatial Domain Assignments
}
