% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAdmixCore-package.R
\docType{package}
\name{cellAdmixCore-package}
\alias{cellAdmixCore}
\alias{cellAdmixCore-package}
\title{cellAdmixCore: R Bindings for the cellAdmix Core}
\description{
R bindings for the C++ cellAdmix core: Xenium and tabular input preparation, cell clustering, persisted NMF fit runs, admixture scoring and correction, admixture auditing, and downstream run inspection helpers.
}
\author{
\strong{Maintainer}: Peter Kharchenko \email{peter.kharchenko@gmail.com}

Authors:
\itemize{
  \item Peter Kharchenko \email{peter.kharchenko@gmail.com}
}

}
\keyword{internal}
